Migration to Javascript Client
Migrated to the well supported Discord.js client. Node seems to be
a more natural fit for this endevor, and client easier to consume.

Recreated all of the commands, and as of now, /sprint works as
expected! There's some missing features that need to be added, but
this is a good starting point.